# Report stats for number of sigverify thread pool threads used