--- a/perf/benches/sigverify.rs
+++ b/perf/benches/sigverify.rs
@@ -31,7 +31,7 @@ fn bench_sigverify_simple(bencher: &mut Bencher) {
     let recycler_out = Recycler::default();
     // verify packets
     bencher.iter(|| {
-        sigverify::ed25519_verify(&mut batches, &recycler, &recycler_out, false, num_packets);
+        sigverify::ed25519_verify(&mut batches, &recycler, &recycler_out, false, num_packets, None);
     })
 }
 
@@ -78,7 +78,7 @@ fn bench_sigverify_uneven(bencher: &mut Bencher) {
     let recycler_out = Recycler::default();
     // verify packets
     bencher.iter(|| {
-        sigverify::ed25519_verify(&mut batches, &recycler, &recycler_out, false, num_packets);
+        sigverify::ed25519_verify(&mut batches, &recycler, &recycler_out, false, num_packets, None);
     })
 }