Adding histogram metrics to understand how many threads in the rayon thread pool are being activated in sigverify to make sense of performance. This will also help ensure #25833 is behaving as expected.

Performance bench data:
Before
```
test bench_sigverify_high_packets_large_batch   ... bench:  33,265,987 ns/iter (+/- 1,381,480)
test bench_sigverify_high_packets_small_batch   ... bench:  36,075,933 ns/iter (+/- 4,672,067)
test bench_sigverify_low_packets_large_batch    ... bench:   4,954,608 ns/iter (+/- 269,857)
test bench_sigverify_low_packets_small_batch    ... bench:   4,937,189 ns/iter (+/- 275,733)
test bench_sigverify_medium_packets_large_batch ... bench:   9,073,593 ns/iter (+/- 2,934,884)
test bench_sigverify_medium_packets_small_batch ... bench:   9,352,987 ns/iter (+/- 1,051,355)
test bench_sigverify_simple                     ... bench:   5,491,902 ns/iter (+/- 813,943)
```
After
```
test bench_sigverify_high_packets_large_batch   ... bench:  33,238,112 ns/iter (+/- 915,765)
test bench_sigverify_high_packets_small_batch   ... bench:  36,729,933 ns/iter (+/- 12,555,219)
test bench_sigverify_low_packets_large_batch    ... bench:   4,947,410 ns/iter (+/- 281,600)
test bench_sigverify_low_packets_small_batch    ... bench:   4,920,535 ns/iter (+/- 178,547)
test bench_sigverify_medium_packets_large_batch ... bench:   8,541,525 ns/iter (+/- 1,071,851)
test bench_sigverify_medium_packets_small_batch ... bench:   9,651,006 ns/iter (+/- 1,203,085)
test bench_sigverify_simple                     ... bench:   5,557,156 ns/iter (+/- 1,161,243)
```