@brennanwatt 2022-06-15
> You can write a test locally which generates some packets and then add some annotations or debug prints to track how many threads were actually used. But I am not sure this is a good use case for metrics; why do we need to gather metrics all the time from the validators how many threads are they using? it seems like a static problem which can be measured locally.

Definitely, I had to test locally extensively during the #25883 dev. I see two main benefits to tracking this at cluster level:
1. Validators can behave differently given the algorithm depends on number/availability of cores at time of verification. Having an idea of how much compute is being allocated towards verification could be useful. That said, we can infer most of this using packet count histogram (I say most because Rayon will not deterministically allocate threads based strictly on packet count).
2. We've discussed making the dedup and verify batch thresholds (`MAX_DEDUP_BATCH` & `MAX_SIGVERIFY_BATCH`) dynamic based on what the validator is actually capable of. However, in order to understand this, we need to understand how much parallelism is actually being deployed. I've slapped together a PoC for this at #25970 to show what I'm talking about

It seems like we can live without 1. 2 is the more compelling argument in my mind. Let me know what you think