@behzadnouri 2022-06-15
> Having an idea of how much compute is being allocated towards verification could be useful.
> However, in order to understand this, we need to understand how much parallelism is actually being deployed.

We already have metrics for `cpu-stats`, `memory-stats`, etc.
You can spin up a gce machine of similar specs to run performance tests.

> We've discussed making the dedup and verify batch thresholds (MAX_DEDUP_BATCH & MAX_SIGVERIFY_BATCH) dynamic based on what the validator is actually capable of.

I am not sure this is a good direction to go.
We spin up a thread pool of the size based on the number of machine cores. After that this is a task scheduling problem which _ideally should be encapsulated in the thread-pool implementation_. The rayon documentation itself for example says that:

> Parallel iterators take care of deciding how to divide your data into tasks; it will **dynamically** adapt for maximum performance.
> https://github.com/rayon-rs/rayon/blob/99b1ff56f/README.md#parallel-iterators-and-more

Adding too many knobs and several layers of dynamic adjustment just makes things too complicated to understand and debug.

If you are seeing evidence that rayon dynamic scheduling is not doing a good enough job and cannot be fine tuned with knobs rayon already provides (`with_min_len`, etc), then perhaps would be better to raise this to rayon developers and see if either they have any suggestion or if their implementation can be improved.

Adding one layer of dynamic adjustment on top of another does not seem to me a good option.