@brennanwatt 2022-06-27
Going to abandon these metrics changes. We will still pursue dynamic verification packet threshold, but we will use a strategy where we only recompute verification throughput when packet count is large and use this as a proxy for max concurrency (i.e. don't actually measure threads in use).